+++ src/dataset.rs
@@ -58,9 +58,9 @@ impl Dataset {
     /// let dataset = Dataset::from_vec(&data, &label, 4).unwrap();
     /// ```
     pub fn from_vec(data: &[f32], labels: &[f32], num_features: i32) -> Result<Self> {
-        let nrows = data.len() as i32 / num_features as i32;
+        let nrows = data.len() as i32 / num_features;
         let ncol = num_features;
-        let is_row_major = 1 as i32; // row-major
+        let is_row_major = 1_i32; // row-major
         let parameters = CString::new("").unwrap();
         let label_name = CString::new("label").unwrap();
 
@@ -81,7 +81,7 @@ impl Dataset {
             handle,
             label_name.as_ptr() as *const c_char,
             labels.as_ptr() as *const c_void,
-            nrows as i32,
+            nrows,
             lightgbm_sys::C_API_DTYPE_FLOAT32 as i32
         ))?;
 
@@ -198,7 +198,8 @@ impl Dataset {
 
         let (m, n) = dataframe.shape();
 
-        let label_series = &dataframe.select_series([label_col_name])?[0].cast(&DataType::Float32)?;
+        let label_series =
+            &dataframe.select_series([label_col_name])?[0].cast(&DataType::Float32)?;
 
         if label_series.null_count() != 0 {
             panic!("Cannot create a dataset with null values, encountered nulls when creating the label array")
